+++ src/patomic/types/align.c
@@ -1,9 +1,35 @@
 #include <patomic/types/align.h>
-#include <patomic/macros/force_inline.h>
+#include <patomic/types/intptr.h>
 
 
-PATOMIC_FORCE_INLINE size_t
+size_t
 patomic_cache_line_size(void)
 {
     return (size_t) PATOMIC_MAX_CACHE_LINE_SIZE;
+}
+
+int
+patomic_align_meets_recommended(
+    const volatile void *ptr,
+    patomic_align_t align
+)
+{
+    patomic_uintptr_t addr = (patomic_uintptr_t) ptr;
+    return (addr % align.recommended == 0);
+}
+
+int
+patomic_align_meets_minimum(
+    const volatile void *ptr,
+    patomic_align_t align,
+    size_t width
+)
+{
+    patomic_uintptr_t addr = (patomic_uintptr_t) ptr;
+    if (addr % align.minimum == 0)
+    {
+        addr %= align.size_within;
+        return (addr + width) <= align.size_within;
+    }
+    else { return 0; }
 }